requirements.txt
streamlit==1.16.0
sentence_transformers==2.2.2
torch==1.13.1
pillow==9.3.0
brotli==1.0.9
pandas==1.5.2
spacy==3.4.3
hydralit_components==1.0.10
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.4.1/en_core_web_lg-3.4.1-py3-none-any.whl